move semantic_segmenter to python/
