+++ python/coral_usb/__init__.py
@@ -4,3 +4,4 @@ from .detector_base import EdgeTPUDetectorBase  # NOQA
 from .face_detector import EdgeTPUFaceDetector  # NOQA
 from .human_pose_estimator import EdgeTPUHumanPoseEstimator  # NOQA
 from .object_detector import EdgeTPUObjectDetector  # NOQA
+from .semantic_segmenter import EdgeTPUSemanticSegmenter  # NOQA
